refactor(post): controller, fegin client 가 공통으로 사용하는 dto에 대해 common으로 분리
