--- cartpost-service/src/main/java/com/example/cartpostservice/commissions/controller/dto/response/PresignedUrlComponent.java
@@ -1,12 +1,0 @@
-package com.example.cartpostservice.commissions.controller.dto.response;
-
-import org.hexagon.core.vo.FileType;
-
-public record PresignedUrlComponent(
-        String key,
-        String queryString,
-        FileType fileType
-) {
-
-}
-
