+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/infra/client/internal/dto/response/DownloadFileComponentResponse.java
@@ -1,5 +1,6 @@
 package com.example.cartpostservice.commissions.infra.client.internal.dto.response;
 
+import com.example.cartpostservice.commissions.common.dto.PresignedUrlComponent;
 import java.util.List;
 
 public record DownloadFileComponentResponse(
